Splitted cylinder series into several files.
