+++ js/Series/Cylinder/CylinderPoint.js
@@ -1,0 +1,60 @@
+/* *
+ *
+ *  Highcharts cylinder - a 3D series
+ *
+ *  (c) 2010-2020 Highsoft AS
+ *
+ *  Author: Kacper Madej
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+'use strict';
+var __extends = (this && this.__extends) || (function () {
+    var extendStatics = function (d, b) {
+        extendStatics = Object.setPrototypeOf ||
+            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
+            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+        return extendStatics(d, b);
+    };
+    return function (d, b) {
+        extendStatics(d, b);
+        function __() { this.constructor = d; }
+        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+    };
+})();
+import BaseSeries from '../../Core/Series/Series.js';
+var ColumnPoint = BaseSeries.seriesTypes.column.prototype.pointClass;
+import U from '../../Core/Utilities.js';
+var extend = U.extend;
+/* *
+ *
+ *  Class
+ *
+ * */
+var CylinderPoint = /** @class */ (function (_super) {
+    __extends(CylinderPoint, _super);
+    function CylinderPoint() {
+        /* *
+         *
+         *  Properties
+         *
+         * */
+        var _this = _super !== null && _super.apply(this, arguments) || this;
+        _this.options = void 0;
+        _this.series = void 0;
+        return _this;
+    }
+    return CylinderPoint;
+}(ColumnPoint));
+extend(CylinderPoint.prototype, {
+    shapeType: 'cylinder'
+});
+/* *
+ *
+ *  Default Export
+ *
+ * */
+export default CylinderPoint;
