+++ ts/Series/Cylinder/CylinderPointOptions.d.ts
@@ -1,0 +1,33 @@
+/* *
+ *
+ *  Highcharts cylinder - a 3D series
+ *
+ *  (c) 2010-2020 Highsoft AS
+ *
+ *  Author: Kacper Madej
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ColumnPointOptions from '../Column/ColumnPointOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface CylinderPointOptions extends ColumnPointOptions {
+    shapeType?: string;
+}
+
+export default CylinderPointOptions;
