+++ ts/Series/Cylinder/CylinderSeriesOptions.d.ts
@@ -1,0 +1,35 @@
+/* *
+ *
+ *  Highcharts cylinder - a 3D series
+ *
+ *  (c) 2010-2020 Highsoft AS
+ *
+ *  Author: Kacper Madej
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ColumnSeriesOptions from '../Column/ColumnSeriesOptions';
+import type CylinderSeries from './CylinderSeries';
+import type { SeriesStatesOptions } from '../../Core/Series/SeriesOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface CylinderSeriesOptions extends ColumnSeriesOptions {
+    states?: SeriesStatesOptions<CylinderSeries>;
+}
+
+export default CylinderSeriesOptions;
